requirements.txt
tensorflow
scikit-learn
spacy
psycopg2
prometheus_client
fpdf
jinja2
flake8
pytest
requests
Faker
Pillow
gitpython
pylint
radon
safety
bandit
coverage
black
